Extract ConversationListPanel
